requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.4.0

# Optional dependencies
librosa>=0.10.0
matplotlib>=3.7.0
tensorboard>=2.13.0
